requirements.txt
dj-database-url
Django==2.2.*
Faker
gunicorn
Pillow
psycopg2
pytz
sqlparse
whitenoise
django-crispy-forms
django-allauth
readtime
